Re-allow jdk11 for sjsonnet
